+++ src/Runtime/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(GameEngine PRIVATE
+  Engine.cpp
+)
